proceding to voting checks the status in the db to see who has rvsped before kicking invitees
